market_monitor/requirements.txt
pandas>=2.2,<3
numpy>=2.1,<3
requests>=2.32,<3
pyarrow>=17,<19
python-dotenv>=1.0,<2
matplotlib>=3.9,<4
statsmodels>=0.14,<1
scikit-learn>=1.5,<2
pytest>=8.3,<9
